BACKEND/pseudonymizationPkcs11-master/requirements.txt
Flask==2.2.2
flask_restx==1.0.3
flask_sqlalchemy==3.0.2
python_pkcs11==0.7.0
okta_jwt_verifier==0.2.3
pandas==1.3.5
python-dotenv==0.21.0
requests==2.28.1
Werkzeug~=2.2.2
flask-cors==3.0.10
redis==4.5.1
pyspark==3.0.0